backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.1
imapclient>=3.0.0
PySocks>=1.7.1
httpx>=0.24.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
